+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes to this project will be documented in this file.
 
 - Function to clean Capacity Tracker care home data by nulling when posts to beds ratio is outside thresholds.
 
+- Model retraining process using Polars and scikit-learn in Fargate
+
 ### Changed
 - Migrated Polars validation scripts over to use PointBlank (compatible with >= Python 3.11), so far:
   - locations_raw